+++ packages/contracts-bedrock/contracts/dispute/FaultDisputeGame.sol
@@ -136,14 +136,12 @@ contract FaultDisputeGame is IFaultDisputeGame, Clone {
         // Compute the position that the claim commits to. Because the parent's position is already
         // known, we can compute the next position by moving left or right depending on whether
         // or not the move is an attack or defense.
-        Position nextPosition = _isAttack
-            ? LibPosition.attack(parent.position)
-            : LibPosition.defend(parent.position);
+        Position nextPosition = _isAttack ? parent.position.attack() : parent.position.defend();
 
         // At the leaf nodes of the game, the only option is to run a step to prove or disprove
         // the above claim. At this depth, the parent claim commits to the state after a single
         // instruction step.
-        if (LibPosition.depth(nextPosition) >= MAX_GAME_DEPTH) {
+        if (nextPosition.depth() >= MAX_GAME_DEPTH) {
             revert GameDepthExceeded();
         }
 
@@ -160,11 +158,11 @@ contract FaultDisputeGame is IFaultDisputeGame, Clone {
         Duration nextDuration = Duration.wrap(
             uint64(
                 // First, fetch the duration of the grandparent claim.
-                Duration.unwrap(LibClock.duration(grandparentClock)) +
+                Duration.unwrap(grandparentClock.duration()) +
                     // Second, add the difference between the current block timestamp and the
                     // parent's clock timestamp.
                     block.timestamp -
-                    Timestamp.unwrap(LibClock.timestamp(parent.clock))
+                    Timestamp.unwrap(parent.clock.timestamp())
             )
         );
 
@@ -178,7 +176,7 @@ contract FaultDisputeGame is IFaultDisputeGame, Clone {
         Clock nextClock = LibClock.wrap(nextDuration, Timestamp.wrap(uint64(block.timestamp)));
 
         // Do not allow for a duplicate claim to be made.
-        ClaimHash claimHash = LibHashing.hashClaimPos(_pivot, nextPosition);
+        ClaimHash claimHash = _pivot.hashClaimPos(nextPosition);
         if (claims[claimHash]) {
             revert ClaimAlreadyExists();
         }
@@ -221,9 +219,56 @@ contract FaultDisputeGame is IFaultDisputeGame, Clone {
      * @inheritdoc IDisputeGame
      */
     function resolve() external returns (GameStatus status_) {
-        // TODO - Resolve the game
-        status = GameStatus.IN_PROGRESS;
-        status_ = status;
+        if (status != GameStatus.IN_PROGRESS) {
+            revert GameNotInProgress();
+        }
+
+        // Search for the left-most dangling non-bottom node
+        // The most recent claim is always a dangling, non-bottom node so we start with that
+        uint256 leftMostIndex = claimData.length - 1;
+        uint256 leftMostTraceIndex = LibPosition.rightIndex(
+            claimData[leftMostIndex].position,
+            MAX_GAME_DEPTH
+        );
+        for (uint256 i = leftMostIndex; i < type(uint64).max; ) {
+            // Fetch the claim at the current index.
+            ClaimData storage claim = claimData[i];
+
+            // Decrement the loop counter; If it underflows, we've reached the root
+            // claim and can stop searching.
+            unchecked {
+                --i;
+            }
+
+            // If the claim is not a dangling node above the bottom of the tree,
+            // we can skip over it. These nodes are not relevant to the game resolution.
+            Position claimPos = claim.position;
+            if (LibPosition.depth(claimPos) == MAX_GAME_DEPTH || claim.countered) {
+                continue;
+            }
+
+            // If the claim is a dangling node, we can check if it is the left-most
+            // dangling node we've come across so far. If it is, we can update the
+            // left-most trace index.
+            uint256 traceIndex = LibPosition.rightIndex(claimPos, MAX_GAME_DEPTH);
+            if (traceIndex < leftMostTraceIndex) {
+                leftMostTraceIndex = traceIndex;
+                leftMostIndex = i + 1;
+            }
+        }
+
+        // If the left-most dangling node is at an even depth, the defender wins.
+        // Otherwise, the challenger wins and the root claim is deemed invalid.
+        // slither-disable-next-line weak-prng
+        if (LibPosition.depth(claimData[leftMostIndex].position) % 2 == 0) {
+            status_ = GameStatus.DEFENDER_WINS;
+        } else {
+            status_ = GameStatus.CHALLENGER_WINS;
+        }
+
+        // Update the game status
+        status = status_;
+        emit Resolved(status_);
     }
 
     /**