+++ packages/replica-healthcheck/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @eth-optimism/replica-healthcheck
 
+## 1.2.5
+
+### Patch Changes
+
+- Updated dependencies [8d7dcc70c]
+- Updated dependencies [d6388be4a]
+  - @eth-optimism/core-utils@0.12.1
+  - @eth-optimism/common-ts@0.8.2
+
 ## 1.2.4
 
 ### Patch Changes